Mark V1 as deprecated, minor codes cleanups
